이미지, 위도, 경도 제외 등록 완료

+++ prj/src/main/java/com/holoyolo/app/trade/service/TradeVO.java
@@ -23,8 +23,8 @@ public class TradeVO {
 	private Date writeDate;
 	@DateTimeFormat(pattern = "yyyy-MM-dd")
 	private Date updateDate;
-	private int latitude;
-	private int longitude;
+	private Double latitude;
+	private Double longitude;
 	private String buyerId;
 	private String sellerId;
